64-bit warning fixes.

+++ win32/SciTEWin.cxx
@@ -434,7 +434,7 @@ FilePath SciTEWin::GetSciteUserHome() {
 // Help command lines contain topic!path
 void SciTEWin::ExecuteOtherHelp(const char *cmd) {
 	GUI::gui_string s = GUI::StringFromUTF8(cmd);
-	unsigned int pos = s.find_first_of('!');
+	size_t pos = s.find_first_of('!');
 	if (pos != GUI::gui_string::npos) {
 		GUI::gui_string topic = s.substr(0, pos);
 		GUI::gui_string path = s.substr(pos+1);
@@ -464,7 +464,7 @@ void SciTEWin::ExecuteHelp(const char *cmd) {
 
 	if (hHH) {
 		GUI::gui_string s = GUI::StringFromUTF8(cmd);
-		unsigned int pos = s.find_first_of('!');
+		size_t pos = s.find_first_of('!');
 		if (pos != GUI::gui_string::npos) {
 			GUI::gui_string topic = s.substr(0, pos);
 			GUI::gui_string path = s.substr(pos + 1);
